♻️ [Refactor]: features/studies/component  features/study-application, widget 계층으로 이동
